+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## 0.0.2
+
+MVP ready
+- add, edit, delete tasks and items
+- create linked tasks
+- reorder tasks
+
 ## 0.0.1
 
 Basic functionality